+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 3.1, 2025-08-06
 ---------------
 - show duplicate hero names with a counter suffix
+- add save-as option to files list context menu
 - fix error on opening savefiles from Restoration of Erathia Steam edition (issue #16)
 - fix showing hero as changed upon opening for some game versions
 - fix file extension filter in files list getting narrowed after deleting file