+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 CHANGELOG
 =========
 
-3.1, 2025-08-06
+3.1, 2025-08-07
 ---------------
 - show duplicate hero names with a counter suffix
 - add save-as option to files list context menu
+- remove files from recent files and heroes lists upon deleting
 - fix error on opening savefiles from Restoration of Erathia Steam edition (issue #16)
 - fix showing hero as changed upon opening for some game versions
 - fix file extension filter in files list getting narrowed after deleting file